pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ttsleaguebot"]

[project]
name = "ttsleaguebot"
version = "0.0.1"
description = "A Table Top Simulator MTG league bot."
authors = [
    {name = "Jacob Lusk", email = "jacobplusk@gmail.com"},
]

readme = "README.md"
keywords = ["mtg", "table top simulator"]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3.10",
]
requires-python = ">=3.10.5"
dependencies = [
    "discord",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib"
]

[project.scripts]
ttsleaguebot    = "ttsleaguebot.main:main"


[project.urls]
Home = ""